Refactor : InputView OutputView 통합
