--- src/main/java/inputView/InputView.java
@@ -1,6 +1,0 @@
-package inputView;
-
-public interface InputView {
-
-  void displayOptions();
-}
