+++ src/main/java/output/ConsoleOutput.java
@@ -23,5 +23,13 @@ public class ConsoleOutput implements Output {
     System.out.println(msg);
   }
 
+  @Override
+  public void displayOptions() {
+    System.out.println("1.조회");
+    System.out.println("2.계산");
+    System.out.println();
+    System.out.print("선택 : ");
+  }
+
 
 }
